Create hibernate panache implementations for repositories

+++ src/main/java/org/example/realworldapi/domain/model/repository/ArticlesUsersRepository.java
@@ -1,14 +1,14 @@
 package org.example.realworldapi.domain.model.repository;
 
-import org.example.realworldapi.domain.model.entity.persistent.ArticlesUsers;
-import org.example.realworldapi.domain.model.entity.persistent.ArticlesUsersKey;
+import org.example.realworldapi.domain.model.entity.ArticlesUsers;
+import org.example.realworldapi.domain.model.entity.ArticlesUsersKey;
 
 import java.util.Optional;
 
 public interface ArticlesUsersRepository {
   boolean isFavorited(Long articleId, Long currentUserId);
 
-  int favoritesCount(Long articleId);
+  long favoritesCount(Long articleId);
 
   ArticlesUsers create(ArticlesUsers articlesUsers);
